% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convGEMs.r
\name{convICD10to9}
\alias{convICD10to9}
\title{Converts ICD10 codes to ICD9 codes}
\usage{
convICD10to9(icd10)
}
\value{
data.table of original codes and converted codes
}
\description{
Converts ICD10 codes to ICD9 codes
}

