% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convGEMs.r
\name{convICD9to10}
\alias{convICD9to10}
\title{Converts ICD9 codes to ICD10 codes}
\usage{
convICD9to10(icd9)
}
\value{
data.table of original codes and converted codes
}
\description{
Converts ICD9 codes to ICD10 codes
}

