% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{getChildrenICD10}
\alias{getChildrenICD10}
\title{Get's all children for a single icd10 code (i.e., those under the
current hierarchy)}
\usage{
getChildrenICD10(icd10)
}
\arguments{
\item{icd10}{icd10 code}
}
\description{
Get's all children for a single icd10 code (i.e., those under the
current hierarchy)
}

