% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{getChildrenICD9}
\alias{getChildrenICD9}
\title{Get's all children for a single icd9 code (i.e., those under the
current hierarchy)}
\usage{
getChildrenICD9(icd9)
}
\arguments{
\item{icd9}{icd9 code}
}
\description{
Get's all children for a single icd9 code (i.e., those under the
current hierarchy)
}

