% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comorbidities.r
\name{getComorbidsICD10}
\alias{getComorbidsICD10}
\title{Get comorbidities for a list of ICD10 codes}
\usage{
getComorbidsICD10(icd10)
}
\arguments{
\item{icd10}{icd10 code}
}
\value{
data.table of icd10 codes and their comorbities
}
\description{
Get comorbidities for a list of ICD10 codes
}

