% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comorbidities.r
\name{getComorbidsICD9}
\alias{getComorbidsICD9}
\title{Get comorbidities for a list of ICD9 codes}
\usage{
getComorbidsICD9(icd9)
}
\arguments{
\item{icd9}{icd9 code}
}
\value{
data.table of icd9 codes and their comorbities
}
\description{
Get comorbidities for a list of ICD9 codes
}

