% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{getICDChildren}
\alias{getICDChildren}
\title{Get's all children for a single icd10 code (i.e., those under the
current hierarchy)}
\usage{
getICDChildren(icd, icdVer)
}
\arguments{
\item{icd}{code}

\item{icdVer}{icd version ("icd9" or "icd10")}
}
\value{
vector of icd codes
}
\description{
Get's all children for a single icd10 code (i.e., those under the
current hierarchy)
}

