% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{getICDRange}
\alias{getICDRange}
\title{Get's all ICD codes within a range (inclusive)}
\usage{
getICDRange(icdSt, icdEnd, icdVer)
}
\arguments{
\item{icdSt}{code start}

\item{icdEnd}{code end}

\item{icdVer}{icd version ("icd9" or "icd10")}
}
\value{
vector of icd codes
}
\description{
Get's all ICD codes within a range (inclusive)
}

