% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{getRangeICD10}
\alias{getRangeICD10}
\title{Get's all ICD-10 codes within a range (inclusive).}
\usage{
getRangeICD10(icd10St, icd10End)
}
\arguments{
\item{icd9St}{start of range}

\item{icd9End}{end of range}
}
\description{
Get's all ICD-10 codes within a range (inclusive).
}

