% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{getRangeICD9}
\alias{getRangeICD9}
\title{Get's all ICD-9 codes within a range (inclusive).}
\usage{
getRangeICD9(icd9St, icd9End)
}
\arguments{
\item{icd9St}{start of range}

\item{icd9End}{end of range}
}
\description{
Get's all ICD-9 codes within a range (inclusive).
}

