% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.r
\docType{data}
\name{icd10Comorbidities}
\alias{icd10Comorbidities}
\title{ICD-10 comorbidity codes for Elixhauser, Charlson, and the
Royal College of Surgeons Charlson score.  Definitions to define
the comorbidities were borrowed from the R comorbidities.icd10 package
at https://github.com/gforge/comorbidities.icd10.}
\format{A data.table with 4589 rows and 4 variables:
\describe{
  \item{icd10}{icd10 code}
  \item{Elixhauser}{Charlson comorbidity based on Deyo and enhanced by Quan}
  \item{Charlsons}{Elixhauser comorbidity enhanced by Quan}
  \item{CharlsonsRCS}{Royal College of Surgeons Charlson}
}}
\source{
\url{ https://github.com/gforge/comorbidities.icd10}
}
\usage{
icd10Comorbidities
}
\description{
ICD-10 comorbidity codes for Elixhauser, Charlson, and the
Royal College of Surgeons Charlson score.  Definitions to define
the comorbidities were borrowed from the R comorbidities.icd10 package
at https://github.com/gforge/comorbidities.icd10.
}
\keyword{datasets}

