% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.r
\docType{data}
\name{icd10Hierarchy}
\alias{icd10Hierarchy}
\title{ICD-10 codes with descriptions and hierarchy (major, chapters, and subchapters), using
same format (column names) as from R's icd9 package:}
\format{A data.table with 91737 rows and 7 variables:
\describe{
  \item{icd10}{icd10 code}
  \item{descShort}{short description}
  \item{descLong}{long description}
  \item{threedigit}{the 3 digit icd10 code}
  \item{major}{description at the level of threedigit}
  \item{subchapter}{second level hierarchy (e.g., A00-A09)}
  \item{chapter}{first level hierarchy (e.g., A00-B99)}
}}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2016-ICD-10-CM-and-GEMs.html}
}
\usage{
icd10Hierarchy
}
\description{
Jack O. Wasey (2015). icd9: Tools for Working with ICD-9 Codes, and
		Finding Comorbidities. R package version 1.2.  https://github.com/jackwasey/icd9
}
\keyword{datasets}

