% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.r
\docType{data}
\name{icd10to9GEMs}
\alias{icd10to9GEMs}
\title{Crosswalk/mappings from CMS to convert between ICD10 and ICD9.  Note: see the
GEMs guide from CMS for appropriate use -- these are not to be considered
direct mappings, but will provide any range from one to many to many to one
mappings.}
\format{A data.table with 79192 rows and 3 variables:
\describe{
  \item{icd9}{icd10 code}
  \item{icd10}{icd9 code}
  \item{flag}{flag describing the type of mapping (see Guide in url GEMs download)}
}}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2016-ICD-10-CM-and-GEMs.html}
}
\usage{
icd10to9GEMs
}
\description{
Crosswalk/mappings from CMS to convert between ICD10 and ICD9.  Note: see the
GEMs guide from CMS for appropriate use -- these are not to be considered
direct mappings, but will provide any range from one to many to many to one
mappings.
}
\keyword{datasets}

