% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.r
\docType{data}
\name{icd9Comorbidities}
\alias{icd9Comorbidities}
\title{ICD-9 comorbidity codes for Elixhauser, Charlson, and the
Royal College of Surgeons Charlson score.  Definitions to define
the comorbidities were borrowed from the R comorbidities.icd10 package
at https://github.com/gforge/comorbidities.icd10.}
\format{A data.table with 2084 rows and 4 variables:
\describe{
  \item{icd9}{icd9 code}
  \item{Elixhauser}{Charlson comorbidity based on Deyo and enhanced by Quan}
  \item{Charlsons}{Elixhauser comorbidity enhanced by Quan}
  \item{CharlsonsRCS}{Royal College of Surgeons Charlson}
}}
\source{
\url{ https://github.com/gforge/comorbidities.icd10}
}
\usage{
icd9Comorbidities
}
\description{
ICD-9 comorbidity codes for Elixhauser, Charlson, and the
Royal College of Surgeons Charlson score.  Definitions to define
the comorbidities were borrowed from the R comorbidities.icd10 package
at https://github.com/gforge/comorbidities.icd10.
}
\keyword{datasets}

