% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.r
\docType{data}
\name{icd9Hierarchy}
\alias{icd9Hierarchy}
\title{ICD-9 codes with descriptions and hierarchy (major, chapters, and subchapters), taken
directly from R's icd9 package:}
\format{A data.table with 17561 rows and 7 variables:
\describe{
  \item{icd9}{icd9 code}
  \item{descShort}{short description}
  \item{descLong}{long description}
  \item{threedigit}{the 3 digit icd9 code}
  \item{major}{description at the level of threedigit}
  \item{subchapter}{second level hierarchy (e.g., 001-009)}
  \item{chapter}{first level hierarchy (e.g., 001-139)}
}}
\source{
\url{https://github.com/jackwasey/icd9}
}
\usage{
icd9Hierarchy
}
\description{
Jack O. Wasey (2015). icd9: Tools for Working with ICD-9 Codes, and
		Finding Comorbidities. R package version 1.2.  https://github.com/jackwasey/icd9
}
\keyword{datasets}

