% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.r
\name{isValidICD}
\alias{isValidICD}
\title{Checks if ICD code(s) are valid.}
\usage{
isValidICD(icd, icdVer)
}
\arguments{
\item{icd}{code(s)}

\item{icdVer}{icd version ("icd9" or "icd10")}
}
\value{
logical(s)
}
\description{
Checks if ICD code(s) are valid.
}

