% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.r
\name{isValidICD10}
\alias{isValidICD10}
\title{Checks if ICD10 code(s) are valid.}
\usage{
isValidICD10(icd10)
}
\arguments{
\item{icd10}{icd10 code}
}
\description{
Checks if ICD10 code(s) are valid.
}

