% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validation.r
\name{isValidICD9}
\alias{isValidICD9}
\title{Checks if ICD9 code(s) are valid.  Values should be
in the correct format, but this method will pad the
string with leading zeros if the length is <3.}
\usage{
isValidICD9(icd9)
}
\arguments{
\item{icd9}{icd9 code}
}
\description{
Checks if ICD9 code(s) are valid.  Values should be
in the correct format, but this method will pad the
string with leading zeros if the length is <3.
}

