% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyrStrWrap.r
\name{s_select}
\alias{s_select}
\title{Modified version of dplyr's select that uses string arguments}
\usage{
s_select(.data, ...)
}
\description{
Modified version of dplyr's select that uses string arguments
}

